requirements.txt
pytest
allure-pytest
requests
playwright==1.47.0
faker
pillow
numpy
